+++ src/aio.c
@@ -198,15 +198,17 @@ static void raio_complete(void *arg) {
 
 }
 
-static void raio_complete_close(void *arg) {
+static void raio_complete_ack(void *arg) {
 
   nano_aio *raio = (nano_aio *) arg;
+  nng_ctx *ctx = (nng_ctx *) raio->data;
   const int res = nng_aio_result(raio->aio);
-  if (res == 0) {
-    nng_msg *msg = nng_aio_get_msg(raio->aio);
-    raio->data = msg;
-    nng_pipe p = nng_msg_get_pipe(msg);
-    nng_pipe_close(p);
+  if (res == 0)
+    raio->data = nng_aio_get_msg(raio->aio);
+  nng_msg *msg;
+  if (nng_msg_alloc(&msg, 0) == 0) {
+    nng_ctx_sendmsg(*ctx, msg, 0);
+    nng_msg_free(msg);
   }
 
 #if NNG_MAJOR_VERSION == 1 && NNG_MINOR_VERSION < 6
@@ -1517,7 +1519,7 @@ SEXP rnng_ncurl_session_close(SEXP session) {
 
 // request ---------------------------------------------------------------------
 
-SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP autoclose, SEXP clo) {
+SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP ack, SEXP clo) {
 
   if (R_ExternalPtrTag(con) != nano_ContextSymbol)
     Rf_error("'context' is not a valid Context");
@@ -1526,7 +1528,7 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
 
   int xc;
   const nng_duration dur = timeout == R_NilValue ? NNG_DURATION_DEFAULT : (nng_duration) Rf_asInteger(timeout);
-  const int ac = LOGICAL(autoclose)[0];
+  const int ac = LOGICAL(ack)[0];
 
   SEXP sendaio, aio, env, fun;
   nano_buf buf;
@@ -1563,8 +1565,9 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
   nano_aio *raio = R_Calloc(1, nano_aio);
   raio->type = RECVAIO;
   raio->mode = nano_matcharg(recvmode);
+  raio->data = ac ? ctx : NULL;
 
-  if ((xc = nng_aio_alloc(&raio->aio, ac ? raio_complete_close : raio_complete, raio))) {
+  if ((xc = nng_aio_alloc(&raio->aio, ac ? raio_complete_ack : raio_complete, raio))) {
     R_Free(raio);
     nng_aio_free(saio->aio);
     R_Free(saio);